fix(checkout): fixed cart context to use product database
