+++ src/components/auth/GenreManagement.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function GenreManagement() {
+  return (
+    <>GenreManagement</>
+  )
+}
+
+export default GenreManagement
