+++ src/components/contexts/CartProvider.jsx
@@ -24,7 +24,7 @@ export const CartProvider = ({ children }) => {
   const addToCart = (
     (itemToAdd) => {
       const isItemInCart = cartItems.some(
-        (item) => item.product_id === itemToAdd.product_id
+        (item) => item._id === itemToAdd._id
       );
 
       if (!isItemInCart) {
@@ -46,9 +46,9 @@ export const CartProvider = ({ children }) => {
 
   const removeFromCart = useCallback(
     (itemId) => {
-      const itemToRemove = cartItems.find((item) => item.product_id === itemId);
+      const itemToRemove = cartItems.find((item) => item._id === itemId);
       const updatedCart = cartItems.filter(
-        (item) => item.product_id !== itemId
+        (item) => item._id !== itemId
       );
       setCartItems(updatedCart);
       if (itemToRemove) {
